download rustlibdir only when necessary

+++ downloadrustlibdir.sh
@@ -1,9 +1,18 @@
 #!/bin/bash
 
-URL="http://static.rust-lang.org/dist/rust-nightly-i686-unknown-linux-gnu.tar.gz"
-TARPATH=rust-nightly-i686-unknown-linux-gnu/lib/rustlib/i686-unknown-linux-gnu/lib/
+TARGET=i686-unknowni-linux-gnu
+URL="http://static.rust-lang.org/dist/rust-nightly-${TARGET}.tar.gz"
+TARPATH=rust-nightly-${TARGET}/lib/rustlib/${TARGET}/lib/
 COMPONENTS=5
 FOLDER=rustlibdir
+
+RUST_TARGET=$(rustc --version verbose | grep host | awk '{print $2}')
+
+if [ "$TARGET" = "$RUST_TARGET" ]
+then
+	echo "Host target is ok. No download needed."
+	exit 0
+fi
 
 rm -rf $FOLDER
 mkdir $FOLDER
